Finalization

+++ app/controllers/api/v1/favorites_controller.rb
@@ -1,8 +1,10 @@
 class Api::V1::FavoritesController < ApplicationController
+
   def create
     user = User.find_by(api_key: params[:api_key])
     if user
       user.favorites.create!(location: params[:location])
+      render json: {data: user.favorites}
     else
       render json: {
         message: "something went wrong!"
@@ -12,15 +14,19 @@ class Api::V1::FavoritesController < ApplicationController
 
   def index
     user = User.find_by(api_key: params[:api_key])
-    render json: user.favorites_forecasts
+    render json: {data: user.favorites_forecasts}
   end
 
   def destroy
     user = User.find_by(api_key: params[:api_key])
     if user
       favorite = user.favorites.find_by(location: params[:location])
-      favorite.destroy
-      render json: user.favorites
+      if favorite
+        favorite.destroy
+        render json: {data: user.favorites}
+      else
+        render json: {message: "That location wasn't found"}
+      end
     else
       render json: {
         message: "something went wrong!"
